Inline 'get_release_and_version' definition
